Move merged menu cache helpers
